Yet more changes for driving
As much as it hurts, i must push out a third commit regarding some minor
tweaks to the robot code.  Rest assured, this is solely tweaking that is
absolutely essential to the robot's function, not a devastating malware
hell-bent on destroying your computer, ravaging your bank accounts, and
causing pure chaos! Cheers!